FIX: false alarms due to wrong print z of gcode paths
The print z was wrong previously when there is a wipe tower, or there
is an object with floating parts.

Jira: STUDIO-2719
udesk: https://bblcs.s5.udesk.cn/entry/ticket/show/2258732

(cherry picked from commit be3097d8044ce9e0079ebf7070b15c7aad7aff0e)

+++ src/libslic3r/GCode/GCodeProcessor.hpp
@@ -97,11 +97,11 @@ namespace Slic3r {
     {
         std::string        _objName1;
         std::string        _objName2;
-        double             _height;
+        float             _height;
         const void *_obj1; // nullptr means wipe tower
         const void *_obj2;
         int                layer = -1;
-        ConflictResult(const std::string &objName1, const std::string &objName2, double height, const void *obj1, const void *obj2)
+        ConflictResult(const std::string &objName1, const std::string &objName2, float height, const void *obj1, const void *obj2)
             : _objName1(objName1), _objName2(objName2), _height(height), _obj1(obj1), _obj2(obj2)
         {}
         ConflictResult() = default;
